Almost posting event page detail properly
So the idea is to pass index from ItemComponent.js into GotoEvent.js and then pass it to EventPageDetail.js
From there we will use the index to run a for loop and only fetch data from specific place. However, I can't seem to figure out how to pass it in properly so This is what I have right now. And also I did some clean up.